moved energy_test.py changes into examples/coreml.py
